Started the process of tuning attitude controller
With only P control, we're still saturating with an VERY oscillatory system. We should look into adding drag for the sake of adding stability to the system. 

As for notes on the response, we have seemingly no control over yaw, with it quickly going unstable. I'm thinking if we add drag that might help us tune the system much more effectively.