Trying to get this closer to SS control rather than PID control
Still working through this. I think I made a bunch of progress (at least on the attitude end of things). The A and B matrix are harder to find than I'd think. Not sure if I can completely remove the omega_dot values from the state vector for the A matrix in the attitude controller.